+++ packages/fuselage/src/components/Select/MultiSelect.js
@@ -11,11 +11,14 @@ const Container = Box.extend('rcx-select', 'div');
 
 const SelectedOptions = React.memo((props) => <Chip {...props}/>);
 
-const prevent = (e) => e.preventDefault() & e.stopPropagation();
+
+const prevent = (e) => e.preventDefault() & e.stopPropagation() & e.nativeEvent.stopImmediatePropagation();
 export const MultiSelect = ({
   value,
   filter,
   options = [],
+  error,
+  disabled,
   anchor: Anchor = Focus,
   onChange = () => {},
   getLabel = ([, label] = []) => label,
@@ -57,15 +60,20 @@ export const MultiSelect = ({
   const ref = useRef();
   const containerRef = useRef();
   return (
-    <Container ref={containerRef} onClick={() => ref.current.focus() & show()}{...props}>
+    <Container className={
+      [
+        error && 'invalid',
+        disabled && 'disabled',
+      ].filter(Boolean).join(' ')
+    } ref={containerRef} onClick={() => ref.current.focus() & show()} disabled={disabled} {...props}>
       <Flex.Item grow={1}>
         <Margins inline='x4'>
           <Flex.Container>
             <Box is='div'>
               <Margins all='neg-x8'>
                 <Chip.Wrapper role='listbox'>
-                  <Anchor ref={ref} aria-haspopup='listbox' onClick={show} onBlur={hide} onKeyUp={handleKeyUp} onKeyDown={handleKeyDown} style={{ order: 1 }} mod-undecorated children={option || placeholder}/>
-                  {currentValue.map((value) => <SelectedOptions role='option' key={value} onMouseDown={(e) => prevent(e) & internalChanged([value]) && false} children={getLabel(options.find(([val]) => val === value))}/>)}
+                  <Anchor disabled={disabled} ref={ref} aria-haspopup='listbox' onClick={show} onBlur={hide} onKeyUp={handleKeyUp} onKeyDown={handleKeyDown} style={{ order: 1 }} mod-undecorated children={!value ? option || placeholder : null}/>
+                  {currentValue.map((value) => <SelectedOptions tabIndex={-1} role='option' key={value} onMouseDown={(e) => prevent(e) & internalChanged([value]) && false} children={getLabel(options.find(([val]) => val === value))}/>)}
                 </Chip.Wrapper>
               </Margins>
             </Box>